refactor : 사용하지 않는 파일 삭제, 모델 파일 통합
