--- docs/project_definition.md
@@ -1,38 +1,0 @@
-# 프로젝트 정의서
-
-## 프로젝트명
-Orblit 1-on-1 AI
-
-## 프로젝트 목적
-1:1 미팅의 품질을 AI를 활용하여 개선하고, 리더와 구성원 간의 효과적인 소통을 지원하는 시스템 구축
-
-## 주요 이해관계자
-- **리더/매니저**: 1:1 미팅을 진행하고 피드백을 받는 주체
-- **HR 담당자**: 조직 내 1:1 미팅 문화 정착 및 품질 관리
-- **구성원**: 1:1 미팅의 상대방
-
-## 프로젝트 범위
-
-### 포함 사항
-1. 음성 파일 STT 변환 기능
-2. 화자 구분 기능
-3. 미팅 템플릿 생성 기능
-4. 대화 분석 및 피드백 생성 기능
-5. 웹 API 제공
-
-### 제외 사항
-1. 실시간 음성 스트리밍 처리
-2. 모바일 애플리케이션
-3. 다국어 지원 (초기 버전은 한국어만 지원)
-
-## 기대 효과
-1. 1:1 미팅 준비 시간 단축 (템플릿 활용)
-2. 미팅 품질 개선을 위한 객관적 피드백 제공
-3. 리더십 코칭 효과
-4. 조직 내 소통 문화 개선
-
-## 성공 지표
-- 음성 파일 STT 변환 정확도 90% 이상
-- 피드백 만족도 80% 이상
-- API 응답 시간 3초 이내
-- 시스템 가용성 99% 이상
