--- docs/requirements_specification.md
@@ -1,74 +1,0 @@
-# 요구사항 명세서
-
-## 1. 기능 요구사항
-
-### 1.1 STT 처리기 (STT Processor)
-
-#### FR-STT-001: 음성 파일 업로드
-- 사용자는 MP3, WAV, M4A, WEBM 형식의 음성 파일을 업로드할 수 있어야 한다
-- 최대 파일 크기는 100MB로 제한한다
-
-#### FR-STT-002: 음성-텍스트 변환
-- OpenAI Whisper API를 사용하여 음성을 텍스트로 변환한다
-- 한국어를 기본으로 지원한다
-
-#### FR-STT-003: 화자 구분
-- 2명 이상의 화자를 구분하여 표시한다
-- 각 발언에 화자 라벨(Speaker 1, Speaker 2 등)을 부여한다
-
-#### FR-STT-004: 변환 결과 저장
-- 변환된 텍스트는 JSON 형식으로 저장한다
-- 타임스탬프와 화자 정보를 포함한다
-
-### 1.2 템플릿 생성기 (Template Generator)
-
-#### FR-TG-001: 템플릿 생성 요청
-- 사용자는 미팅 목적을 텍스트로 입력할 수 있다
-- 추가 컨텍스트 정보를 선택적으로 입력할 수 있다
-
-#### FR-TG-002: AI 기반 템플릿 생성
-- GPT-4를 사용하여 상황에 맞는 질문 템플릿을 생성한다
-- 5-10개의 핵심 질문을 포함한다
-
-#### FR-TG-003: 템플릿 카테고리
-- 신규 입사자 온보딩
-- 성과 리뷰
-- 경력 개발
-- 프로젝트 체크인
-- 일반 체크인
-
-### 1.3 1:1 피드백 생성기 (1-on-1 Feedback Generator)
-
-#### FR-FB-001: 대화 분석
-- STT 변환된 텍스트를 입력받아 분석한다
-- 발언 비율, 질문 횟수, 대화 턴 등을 계산한다
-
-#### FR-FB-002: 피드백 생성
-- GPT-4를 사용하여 개선 피드백을 생성한다
-- 정량적 지표와 정성적 조언을 포함한다
-
-#### FR-FB-003: 피드백 영역
-- 경청 스킬
-- 질문의 질
-- 대화 균형
-- 감정적 지원
-- 액션 아이템 도출
-
-## 2. 비기능 요구사항
-
-### 2.1 성능 요구사항
-- NFR-001: API 응답 시간은 95%의 요청에서 3초 이내여야 한다
-- NFR-002: 동시 사용자 100명을 지원해야 한다
-
-### 2.2 보안 요구사항
-- NFR-003: 모든 API 통신은 HTTPS를 사용한다
-- NFR-004: 업로드된 음성 파일은 암호화하여 저장한다
-- NFR-005: API 키는 환경 변수로 관리한다
-
-### 2.3 사용성 요구사항
-- NFR-006: RESTful API 설계 원칙을 따른다
-- NFR-007: 명확한 에러 메시지를 제공한다
-
-### 2.4 확장성 요구사항
-- NFR-008: 마이크로서비스 아키텍처로 설계한다
-- NFR-009: 새로운 언어 추가가 용이해야 한다
