+++ pyproject.toml
@@ -5,17 +5,14 @@ description = "AI-powered 1-on-1 meeting feedback system"
 authors = ["Your Name"]
 readme = "README.md"
 packages = [{include = "src"}]
-packages = [{include = "src"}]
 
 [tool.poetry.dependencies]
-python = "^3.11"
 python = "^3.11"
 fastapi = "^0.104.0"
 uvicorn = {extras = ["standard"], version = "^0.24.0"}
 openai = "^1.0.0"
 pydantic = "^2.9.0"
 # pydantic-settings = "^2.0.0"
-pydantic = "^2.9.0"
 python-multipart = "^0.0.6"
 python-dotenv = "^1.0.0"
 aiofiles = "^23.2.0"
@@ -30,10 +27,6 @@ langchain-openai = "^0.2.0"
 langchain-google-vertexai = "^2.0.28"
 requests = "^2.32.4"
 streamlit = "^1.48.0"
-langchain = "^0.3.0"
-langchain-google-vertexai = "^2.0.28"
-requests = "^2.32.4"
-streamlit = "^1.48.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.4.0"
@@ -43,13 +36,6 @@ black = "^23.9.0"
 flake8 = "^6.1.0"
 mypy = "^1.5.0"
 pre-commit = "^3.4.0"
-
-
-[tool.pytest.ini_options]
-pythonpath = ["."]
-testpaths = ["tests"]
-python_files = ["test_*.py"]
-python_functions = ["test_*"]
 
 [tool.pytest.ini_options]
 pythonpath = ["."]
