--- src/models/recording.py
@@ -1,1 +1,0 @@
-# 녹음 관련 모델
