--- src/models/stt_llm_analysis.py
@@ -1,1 +1,0 @@
-# STT→LLM 분석 관련 모델
