--- src/models/template.py
@@ -1,1 +1,0 @@
-# 템플릿 생성 관련 모델
