--- src/prompts/stt_llm_prompts.py
@@ -1,1 +1,0 @@
-# STT→LLM 분석용 프롬프트
