--- src/prompts/template_prompts.py
@@ -1,1 +1,0 @@
-# 템플릿 생성용 프롬프트
